+++ WrikeExplorer/Public/Get-WrikeContact.ps1
@@ -1,6 +1,14 @@
 function Get-WrikeContact {
     [CmdletBinding()]
+    [OutputType([WrikeContact])]
     param (
+        # Specifies one or more optional ContactId values for known contacts
+        [Parameter(ValueFromPipelineByPropertyName)]
+        [ValidateNotNullOrEmpty()]
+        [Alias('Id')]
+        [string[]]
+        $ContactId,
+
         # Specifies that the Wrike Contact record for the current user should be returned
         [Parameter()]
         [Alias('CurrentUser')]
@@ -20,6 +28,10 @@ function Get-WrikeContact {
     )
 
     process {
+        if ($ContactId.Count -gt 100) {
+            Write-Error 'The Wrike API imposes a limit of 100 contact IDs in a contact history query.'
+            return
+        }
         $query = @{}
         if ($MyInvocation.BoundParameters.ContainsKey('Me')) {
             $query.me = $Me.ToString().ToLower()
@@ -28,12 +40,18 @@ function Get-WrikeContact {
             $query.deleted = $Deleted.ToString().ToLower()
         }
         if ($MyInvocation.BoundParameters.ContainsKey('Include')) {
+            # De-duplicate the list of optional fields
+            $Include = $Include | Group-Object | Select-Object -ExpandProperty Name
             $json = $Include | ConvertTo-Json -Compress
             if ($Include.Count -lt 2) {
                 $json = "[$json]"
             }
             $query.fields = $json
         }
-        Invoke-WrikeApi -Path contacts -ResponseType contacts -Query $query
+        $path = "contacts"
+        if ($null -ne $ContactId -and $ContactId.Count -gt 0) {
+            $path += '/' + [string]::Join(',', $ContactId)
+        }
+        Invoke-WrikeApi -Path $path -ResponseType contacts -Query $query
     }
 }