+++ CHANGELOG.md
@@ -6,6 +6,12 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [0.1.1] - unreleased
+### Added
+- New Get-WrikeContactHistory function.
+- Formatting for Wrike contact records to improve output in terminal or Out-GridView for example.
+### Changed
+- Added a ContactId parameter to Get-WrikeContact with an alias for Id.
+- Deduplicate the Include parameter values in case user provides the same optional field names more than once.
 
 ## [0.1.0] - 2021-09-20